chore: rename folder to traits
